Bug fixes

See merge request adviqotech/adviser-app!39

+++ lib/presentation/screens/advisor_preview/advisor_preview_cubit.dart
@@ -33,7 +33,7 @@ class AdvisorPreviewCubit extends Cubit<AdvisorPreviewState> {
 
   Map<String, dynamic> getSelectedLanguageDetails(MarketsType language) {
     Map<String, dynamic> details = {};
-    details[ratingKey] = userProfile.rating?[language] as double;
+    details[ratingKey] = userProfile.rating?[language];
     details[titleKey] = (userProfile.localizedProperties
                 ?.toJson()[language.name] as PropertyByLanguage?)
             ?.statusMessage ??
